update:增加neo4j的model和repository以及neo4j配置

+++ backend/src/main/java/edu/fudan/se/sctap_lowcode_tool/neo4jrepository/DeviceNodeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.neo4j.repository.Neo4jRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DeviceNodeRepository extends Neo4jRepository<DeviceNode, Integer> {
+public interface DeviceNodeRepository extends Neo4jRepository<DeviceNode, Long> {
 }
