+++ backend/src/main/java/edu/fudan/se/sctap_lowcode_tool/neo4jrepository/PersonNodeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.neo4j.repository.Neo4jRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PersonNodeRepository extends Neo4jRepository<PersonNode, Integer> {
+public interface PersonNodeRepository extends Neo4jRepository<PersonNode, Long> {
 }
