+++ backend/src/main/java/edu/fudan/se/sctap_lowcode_tool/neo4jrepository/SpaceNodeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.neo4j.repository.Neo4jRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SpaceNodeRepository extends Neo4jRepository<SpaceNode, Integer> {
+public interface SpaceNodeRepository extends Neo4jRepository<SpaceNode, Long> {
 }
