+++ backend/src/main/java/edu/fudan/se/sctap_lowcode_tool/neo4jrepository/StateNodeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.neo4j.repository.Neo4jRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StateNodeRepository extends Neo4jRepository<StateNode, Integer> {
+public interface StateNodeRepository extends Neo4jRepository<StateNode, Long> {
 }
